fix(orchestrator): address review on waitForSessionReady
- Replace time.After(pollInterval) in the poll loop with a time.NewTicker +
  defer Stop, so a long wait doesn't pile up ~maxWait/pollInterval live timers
  (github-actions review). Mirrors waitForAgentPromptReady.
- Move TestWaitForSessionReady_BoundsQueriesWithNonCancellableContext +
  deadlineRecordingRepo out of task_operations_test.go into
  task_operations_wait_test.go — appending them pushed that file past the
  800-line revive file-length-limit, failing changed-file golangci-lint.
- Register t.Cleanup(cancel) on the cancelable context in
  TestEnsureSessionRunning_SurvivesCallerContextCancellationDuringResume
  (coderabbit).
